requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.0
sqlparse==0.5.5
